requirements.txt
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
phonenumbers==8.13.28
python-dotenv==1.0.0
Bootstrap-Flask==2.3.3
boto3==1.34.27
botocore==1.34.27
requests
pytz==2023.3
twilio==8.11.1
APScheduler==3.10.4